add Date & Time to each marker popup
I wanted to show the date and time of the quake on the popup - it seems
like important information and can also highlight how the data is pulled
whenever the site is accessed or refreshed. I used the New Date()
constructor to create a Date object, which has a built-in string
representation and can be used in the bindPopup() funciton.